# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS API
  version: 0.1.0
  description: OpenAPI generado manualmente basado en comportamiento actual.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
      example:
        message: Registro no encontrado.
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
      required: [page, limit, total, total_pages]
      example:
        page: 1
        limit: 20
        total: 1
        total_pages: 1
...
    Carrera:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        is_active: { type: boolean }
      required: [id, name, is_active]
      example:
        id: 1
        name: Ingenieria
        is_active: true
...
    Attempt:
      type: object
      properties:
        id: { type: integer }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        score: { type: number, nullable: true }
        quiz:
          type: object
          properties:
            id: { type: integer }
            title: { type: string }
        user:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            first_name: { type: string }
            last_name: { type: string }
      required: [id, started_at, quiz, user]
      example:
        id: 3
        started_at: '2026-02-05 06:23:53'
        finished_at: '2026-02-05 06:23:54'
        score: 100
        quiz:
          id: 3
          title: Quiz 1
        user:
          id: 30
          email: student@lms.local
          first_name: Student
          last_name: One
...
paths:
...
  /structure/carreras:
    get:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: is_active
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Carrera' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Structure]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                is_active: { type: boolean }
              required: [name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Carrera' }
  /structure/carreras/{id}:
    put:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                is_active: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Carrera' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
    delete:
      tags: [Structure]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...
  /assessments/attempts:
    get:
      tags: [Assessments]
      security: [bearerAuth: []]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: quiz_id
          schema: { type: integer }
        - in: query
          name: user_id
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Attempt' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Assessments]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quiz_id: { type: integer }
              required: [quiz_id]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Attempt' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /assessments/attempts/{id}/finish:
    post:
      tags: [Assessments]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Attempt' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /assessments/attempts/{id}:
    delete:
      tags: [Assessments]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...